Fix(hypothetical): avoid race condition
in renderWait(), it was possible for the render exit() promise to resolve b4 the callback finishes executing (this is because we called ender() b4 cb() to avoid the double console.log issue).

now we forward the args to the callback once we are certain the renderer is finished, which is a more sequential solution. as a bonus we directly return the cb() result